backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
transformers==4.35.0
torch==2.1.0
google-generativeai==0.3.0
python-dotenv==1.0.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
gunicorn==21.2.0
pydantic==2.5.0
bleach==6.1.0